+++ src/services/api.js
@@ -44,4 +44,3 @@ var Api = (function() {
 })();
 
 module.exports = Api;
-